Fix disable hotkeys option and double presses
Replace getUInt() with getBool() for "browser/disable_hotkeys" since
getUInt() seems to always returns a falsey value.

Additionally, fix a bug where key press events that are not directly
handled by a DOM object are sent to the webpage twice. This happens in
the case where the WebView is focused, and the key event is sent from
the WebView to the webpage, and then sent to the MainWindow object, and
then sent back to the WebView. This causes problems for javascript key
event listeners. This was fixed by removing the MainWindow's behavior of
propogating key presses to the WebView, and setting the focus behavior
of the MainWindow to NoFocus.